Worked on cleanup process

+++ src/main/java/com/zggis/dobby/services/MediaServiceImpl.java
@@ -2,11 +2,17 @@ package com.zggis.dobby.services;
 
 import java.io.File;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+import com.zggis.dobby.batch.ConsoleColor;
+
 @Component
 public class MediaServiceImpl {
+
+	private static final Logger logger = LoggerFactory.getLogger(MediaServiceImpl.class);
 
 	@Value("${media.dir}")
 	private String mediaDir;
@@ -38,8 +44,15 @@ public class MediaServiceImpl {
 
 	public void deleteTempDirectory() {
 		if (cleanup) {
+			logger.info("Cleaning up TEMP directory...");
 			deleteDirectory(tempDir);
+		} else {
+			logger.info("Cleanup is disabled by configuration.");
 		}
+	}
+
+	public boolean isCleanup() {
+		return cleanup;
 	}
 
 	private static void createDirectory(String directoryStr) {
@@ -49,16 +62,22 @@ public class MediaServiceImpl {
 		}
 	}
 
-	private static void deleteDirectory(String directoryName) {
+	private void deleteDirectory(String directoryName) {
 		File directory = new File(directoryName);
+		boolean dirEmpty = true;
 		if (directory.exists()) {
 			File[] directoryListing = directory.listFiles();
 			if (directoryListing != null) {
 				for (File file : directoryListing) {
-					file.delete();
+					logger.warn(ConsoleColor.YELLOW.value
+							+ "This file {} was found in the TEMP directory {}, but Dobby did not put it there. It will not be touched."
+							+ ConsoleColor.NONE.value, file.getName(), tempDir);
+					dirEmpty = false;
 				}
 			}
-			directory.delete();
+			if (dirEmpty) {
+				directory.delete();
+			}
 		}
 	}
 }
